Extract `GreaterThan` tests into separate test case
